+++ cliente1/keycloak-frontend/src/app/app.component.html
@@ -1,2 +1,2 @@
-<app-menu></app-menu>
+<app-menu [isLogged]="isLogged" [isRoot]="isRoot"></app-menu>
 <router-outlet></router-outlet>
