+++ cliente1/keycloak-frontend/src/app/menu/menu.component.html
@@ -10,17 +10,17 @@
       <li class="nav-item ">
         <a class="nav-link" href="#">Home <span class="sr-only">(current)</span></a>
       </li>
-      <li class="nav-item ">
-        <a class="nav-link" href="#">Action</a>
+      <li *ngIf="isLogged" class="nav-item ">
+        <a  class="nav-link" href="#">Vista</a>
       </li>
-      <li class="nav-item">
+      <li *ngIf="isRoot" class="nav-item">
         <a class="nav-link " href="#">Admin</a>
       </li>
-      <li class="nav-item">
+      <li  *ngIf="isRoot" class="nav-item">
         <a class="nav-link " href="#">Nuevo</a>
       </li>
     </ul>
-    <button class="btn btn-outline-warning" (click)="login()">Login</button>
-    <button class="btn btn-outline-warning" (click)="logout">Logout</button>
+    <button *ngIf="!isLogged"  class="btn btn-outline-warning" (click)="login()">Login</button>
+    <button  *ngIf="isLogged" class="btn btn-outline-warning" (click)="logout">Logout</button>
   </div>
 </nav>
