+++ cliente1/keycloak-frontend/src/app/services/login.service.ts
@@ -12,7 +12,6 @@ export class LoginService {
 
   public login(): void{
     this.oauthService.initImplicitFlowInternal();
-
   }
 
   public logut(): void{
@@ -20,5 +19,19 @@ export class LoginService {
     this.oauthService.logOut();
     // keycloak no soporta el revoke token ouath
   }
+
+  public getIsLogged(): boolean{
+    return (this.oauthService.hasValidIdToken() && this.oauthService.hasValidAccessToken());
+  }
+
+  public getAdmin(): boolean{
+    const t = this.oauthService.getAccessToken();
+    const payLoad = t.split('.')[1];
+    const decodeJson = atob(payLoad);
+    const decode = JSON.parse(decodeJson);
+    return decode.realm_acces.roles.indexOf('real-admin') !== -1 ;
+  }
+
+
 
 }
